# specs/openapi.yaml
openapi: 3.1.0
info:
  title: Writ API
  version: 0.1.0
  summary: Governed methodology, evidence, evaluation, and release API.
  description: |
    Planning contract. Command endpoints perform explicit state transitions and
    require authorization, optimistic concurrency, idempotency, and audit events.
servers:
  - url: http://localhost:3000
    description: Local development
security:
  - oidc: []
...
paths:
...
  /v1/claims:
    get:
      tags: [evidence]
      operationId: listClaims
      parameters:
        - $ref: "#/components/parameters/Cursor"
        - $ref: "#/components/parameters/Limit"
        - name: status
          in: query
          schema: {type: string}
        - name: subject_ref
          in: query
          schema: {type: string}
      responses:
        "200": {$ref: "#/components/responses/GenericPage"}
    post:
      tags: [evidence]
      operationId: createClaim
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "evidence.schema.json#/$defs/claim"
      responses:
        "201":
          description: Candidate claim created.
          content:
            application/json:
              schema:
                $ref: "evidence.schema.json#/$defs/claim"
        "409": {$ref: "#/components/responses/Problem"}
...
  /v1/actions:
    get:
      tags: [evidence]
      operationId: listActions
      parameters:
        - $ref: "#/components/parameters/Cursor"
        - $ref: "#/components/parameters/Limit"
        - name: jurisdiction
          in: query
          schema: {type: string}
        - name: status
          in: query
          schema: {type: string}
      responses:
        "200": {$ref: "#/components/responses/GenericPage"}
    post:
      tags: [evidence]
      operationId: createAction
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "evidence.schema.json#/$defs/action"
      responses:
        "201":
          description: Candidate action created.
          content:
            application/json:
              schema:
                $ref: "evidence.schema.json#/$defs/action"
        "409": {$ref: "#/components/responses/Problem"}
...
components:
  securitySchemes:
    oidc:
      type: openIdConnect
      openIdConnectUrl: https://identity.example/.well-known/openid-configuration
  parameters:
    Cursor:
      name: cursor
      in: query
      schema: {type: string}
    Limit:
      name: limit
      in: query
      schema: {type: integer, minimum: 1, maximum: 200, default: 50}
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema: {type: string, minLength: 8, maxLength: 200}
...
  schemas:
...
    Problem:
      type: object
      additionalProperties: false
      required: [type, title, status, code, detail, request_id]
      properties:
        type: {type: string, format: uri}
        title: {type: string}
        status: {type: integer, minimum: 400, maximum: 599}
        code: {type: string}
        detail: {type: string}
        instance: {type: string}
        request_id: {type: string}
        context: {type: object}
    Page:
      type: object
      required: [items]
      properties:
        items: {type: array, items: {type: object}}
        next_cursor: {type: string}
        request_id: {type: string}
  responses:
    Problem:
      description: Problem details.
      content:
        application/problem+json:
          schema: {$ref: "#/components/schemas/Problem"}
...
    GenericPage:
      description: Cursor-paginated resources.
      content:
        application/json:
          schema: {$ref: "#/components/schemas/Page"}
...